--- src/components/IndividualWebinar/Speakers/index.module.css
@@ -1,56 +1,0 @@
-.layout {
-  padding-block: 6rem;
-}
-.hr {
-  margin-block: 1.5rem;
-}
-.cardsContainer {
-  display: grid;
-  gap: 1.5rem;
-  grid-template-columns: repeat(1, 1fr);
-}
-@media screen and (min-width: 800px) {
-  .cardsContainer {
-    grid-template-columns: repeat(2, 1fr);
-  }
-}
-
-@media screen and (min-width: 1200px) {
-  .cardsContainer {
-    grid-template-columns: repeat(3, 1fr);
-  }
-}
-.card {
-  border: 1px solid #dee4ed;
-  border-radius: 8px 8px 8px 8px;
-  padding: 3.5rem 0;
-  overflow: hidden;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-}
-
-.card h3 {
-  font-size: 24px;
-  padding-bottom: 10px;
-  margin-top: 1rem;
-}
-
-.card img {
-  width: 15rem;
-  height: 15rem;
-  display: block;
-  border-radius: 9999px;
-  overflow: hidden;
-}
-
-.card p {
-  font-size: 18px;
-  color: #ffffff;
-  line-height: 145%;
-}
-
-.btn1 {
-  margin-top: 1rem;
-}
